Bridge React tracer via postMessage and add debug logs
